Move out vg version
